Se implemento GSON y se verifica que cliente exista

+++ backend/src/main/java/opActAcct/starBankbackend/repository/JsonRepository/CurrentAccountJSON.java
@@ -3,17 +3,17 @@ package opActAcct.starBankbackend.repository.JsonRepository;
 import com.google.gson.Gson;
 import com.google.gson.internal.LinkedTreeMap;
 import opActAcct.starBankbackend.model.CurrentAccount;
+import opActAcct.starBankbackend.model.SavingAccount;
 import opActAcct.starBankbackend.repository.ICurrentAccountRepository;
-import opActAcct.starBankbackend.repository.IMap;
 import opActAcct.starBankbackend.repository.exception.DuplicateKeyException;
 import java.io.*;
 import java.util.*;
 
 
-public class CurrentAccountJSON implements ICurrentAccountRepository, IJson, IMap{
+public class CurrentAccountJSON implements ICurrentAccountRepository, IJson{
 
-    private String fileName = "current_accounts.json";
-    private HashMap<String, LinkedTreeMap<String,Object>> accounts= new HashMap<>();
+    private static String fileName = "current_accounts.json";
+    private static HashMap<String, LinkedTreeMap<String,Object>> accounts= new HashMap<>();
 
     public CurrentAccountJSON() {
         try (BufferedReader br = new BufferedReader(new FileReader(fileName))){
@@ -26,7 +26,7 @@ public class CurrentAccountJSON implements ICurrentAccountRepository, IJson, IMa
             System.out.println("Problema, mirar error IOE");
         }
     }
-
+/*
     @Override
     public String mapToStringInJSONFormat(String key, LinkedTreeMap map) {
         String stringTransactions = "[\n";
@@ -64,8 +64,8 @@ public class CurrentAccountJSON implements ICurrentAccountRepository, IJson, IMa
 
         return mapAttributes;
     }
+    */
 
-    @Override
     public Object readJson() {
 
         Gson gson = new Gson();        //Variable para API Gson. Esta nos regalará los metodos para pasar entre JAVA y Json
@@ -102,9 +102,40 @@ public class CurrentAccountJSON implements ICurrentAccountRepository, IJson, IMa
     }
 
     @Override
-    public void writeJson(String string) {
+    public void writeJson(Object objectToWrite) throws DuplicateKeyException {
+
+        /* (1)
+         * Se comprueba que el ID no exista.
+         */
+        CurrentAccount accountToWrite = (CurrentAccount) objectToWrite;
+        accounts= (HashMap) readJson();
+        if(accounts.containsKey( accountToWrite.getAccount_id())){     //Si ingresa es porque el id ya existe
+            throw new DuplicateKeyException(accountToWrite.getAccount_id());
+        }
+
+        /* (2)
+         * Construye el Mapa Organizado con los atributos del objeto.
+         */
+        LinkedTreeMap mapAttributes = new LinkedTreeMap();
+        mapAttributes.put("isActive", accountToWrite.isStatus());
+        mapAttributes.put("client_id", accountToWrite.getClient_id());
+        mapAttributes.put("balance",accountToWrite.getBalance());
+        mapAttributes.put("sucursal_id", accountToWrite.getSucursal_id());
+        mapAttributes.put("transactions", accountToWrite.getTransactions());
+        mapAttributes.put("creation_date", accountToWrite.getCreation_date());
+
+        accounts.put(accountToWrite.getAccount_id(), mapAttributes);
+
+        /* (3)
+         *
+         */
+        Gson gson = new Gson();
+        String stringToWrite = gson.toJson(accounts);
+
+
+        //copiar en el archivo
         try (BufferedWriter bw = new BufferedWriter(new FileWriter(fileName))) {
-            bw.write(string);
+            bw.write(stringToWrite);
         } catch (FileNotFoundException ex) {
             System.out.println(ex.getMessage());
         } catch (IOException ex) {
@@ -112,56 +143,12 @@ public class CurrentAccountJSON implements ICurrentAccountRepository, IJson, IMa
         }
     }
 
-    @Override
-    public void addJson(Object object) throws DuplicateKeyException {
-        /**
-         * (1)
-         * Se comprueba que el ID no exista.
-         */
-        CurrentAccount account = (CurrentAccount) object;
-        accounts= (HashMap) readJson();
-        if(accounts.containsKey( account.getAccount_id())){     //Si ingresa es porque el id ya existe
-            throw new DuplicateKeyException(account.getAccount_id());
-        }
-
-        /**
-         * (2)
-         * Agrega al diccionario de los clientes, el nuevo objeto tipo HashMap con los datos del cliente
-         */
-        accounts.put(((CurrentAccount) object).getAccount_id(), objectToMapInJSONFormat(object));
-
-        /**
-         * (3)
-         * Se crea el String (de todos los clientes anidados) para poner en el JSON.
-         */
-        String stringAccountHashMapFormat= "{\n";
-        //Se Itera el Diccionario
-        for (Map.Entry<String, LinkedTreeMap<String,Object>> accountLoop : accounts.entrySet()) {
-            //Cada objeto del diccionario se convierte en string. Esto con el método mapToStringInJSONFormat()
-            stringAccountHashMapFormat = stringAccountHashMapFormat.concat(
-                    mapToStringInJSONFormat(accountLoop.getKey() ,accountLoop.getValue()) + ",\n");
-
-        }
-        //Elimina la última coma (,) del string que contiene todos los clientes
-        stringAccountHashMapFormat = stringAccountHashMapFormat.substring(0,
-                stringAccountHashMapFormat.lastIndexOf(','));
-        //Añade el último (}) para el string que contiene todos los clientes
-        stringAccountHashMapFormat = stringAccountHashMapFormat.concat("\n}");
-
-        /**
-         * (4)
-         * Se copia sobre el archivo el string con todos las cuentas en clase JSON.HashMap
-         */
-        writeJson(stringAccountHashMapFormat);
-    }
-
-    @Override
     public void createANewAccount(String account_id,String client_id, String sucursal_id) throws DuplicateKeyException {
         if(accounts.containsKey( account_id )){     //Si ingresa es porque el id ya existe
             throw new DuplicateKeyException(account_id);
         }
         Date fecha = new Date();
         CurrentAccount account = new CurrentAccount(account_id, client_id, sucursal_id, false, Float.parseFloat("0"), new ArrayList(), fecha.toString());
-        addJson(account);
+        writeJson(account);
     }
 }
